Move the five reusable shapes into src/components/ui/
`ui/` first because it is the sink: all 14 cross-boundary relative
imports in `src/components/` point into it and none point out, so moving
it flushes every crossing out while the rest of the tree is still where
it was. Nothing left to discover in the later slices.

`touchTarget` comes with them. It is presentational, imports nothing from
`src/lib`, names nothing about the program, the session, the interest
list or the conditions, and is the only module all three directories
consume. Leaving it at the root would have made the root a shared-utility
bucket, which contradicts the root-is-the-default rule in ADR-0018.

Nine files move with no content change at all — the nine internal imports
were already ui-to-ui. The 14 relative and 9 alias specifiers that named
them are the whole diff.
