+++ src/components/ui/ReservedSlot.tsx
