+++ src/components/ui/SnapSection.test.tsx
