+++ src/components/ui/SnapSection.tsx
